refactor: 로직 분리 (FE-107)
